pyproject.toml
[project]
name = "ucf-desktop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.21.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pdfplumber>=0.11.0",
    "pillow>=10.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.0.0",
    "pyxlsb>=1.0.10",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
]
